Add files via upload
Clone the Lesson-02 branch in Github.  Will be working with SleepData.
https://github.com/jgrissom/NETDBProgramming/tree/Lesson-02
The program currently prompts a user if they want to 1) create or 2) parse a data file. The data creation portion of the program is already complete. The user is asked how many full weeks of data is desired and it generates data for that number of weeks and saves it to a text file. The data represents the number of hours of sleep per night for a user-provided number of weeks.
Your assignment is to complete the TODO in the code. You will need to parse the file and display a ”weekly report” displayed at the system console. You should use string interpolation to complete the task. Start by adjusting the file variable on line #18 to match your file system.

Of course, you will need to figure how the data is formatted in the text file that is generated. To get you started, each line, represents 1 weeks worth of data. The data contains the start date of the week followed by the # of hours of slept for each night (Sunday – Saturday).
There is a “,” separating the date from the hours of sleep and a “|” is used as a delimiter for each night’s hours of sleep.
The next slide demonstrates how the report should look. The screen capture on the left is your minimal goal. For double-secret, extra-credit, create the screen on the right.
In order to receive the double-secret, extra-credit, your report must be formatted EXACTLY like the example shown.